requirements.txt
langchain==0.3.3
langchain_google_genai==2.0.1
Pillow==11.0.0
protobuf==5.28.2
python-dotenv==1.0.1
streamlit==1.39.0
~treamlit==1.38.0
